requirements.txt
google-generativeai
pandas
nltk
matplotlib
seaborn
scipy
